Remove ablity to compile on OpenBSD
The check for OpenBSD cause make to fail on debain with the error:

checking for TLS... no
configure: error: in `/home/rebecca/erltls':
configure: error: The pkg-config script could not be found or is too old.  Make
sure it
is in your PATH or set the PKG_CONFIG environment variable to the full
path to pkg-config.

Alternatively, you may set the environment variables TLS_CFLAGS
and TLS_LIBS to avoid the need to call pkg-config.
See the pkg-config man page for more details.

To get pkg-config, see <http://pkg-config.freedesktop.org/>.
See `config.log' for more details