Add a sample client in C
This way we can write an Erlang program that acts just like it.

Connect to mike-burns.com:443 and print the result to stdout.

There might be/probably is a memory leak in here (maybe in
from_server()), but I don't think any of the TLS data are leaking so we
can ignore it for now.