Some sample C client cleanups
It seems that the branch in which I was hacking on a C client got
merged, and messily, so allow me to commit some improvements to that and
intentionally merge this branch.

Only some of the history is being re-written here because of the
surprise merge and a desire not to futz with git too much.

The improvements:

- Only pass one cipher to tls_config_set_ciphers(3), for simplicity and
  readability.
- Portable systems might not have a hardcoded path to the system cert
  files, so allow people to pass it manually via the command line.